SiteUser 엔티티 추가, 다른 엔티티의 주키를 Long 타입으로

+++ src/main/java/com/ll/exam/sbb/answer/AnswerController.java
@@ -20,10 +20,11 @@ public class AnswerController {
     private final AnswerService answerService;
 
     @PostMapping("/create/{id}")
-    public String detail(Model model, @PathVariable int id, @Valid AnswerForm answerForm, BindingResult bindingResult) {
+    public String detail(Model model, @PathVariable long id, @Valid AnswerForm answerForm, BindingResult bindingResult) {
+
         Question question = this.questionService.getQuestion(id);
 
-        if(bindingResult.hasErrors()){
+        if ( bindingResult.hasErrors() ) {
             model.addAttribute("question", question);
             return "question_detail";
         }
