+++ src/main/java/com/ll/exam/sbb/answer/AnswerRepository.java
@@ -1,13 +1,12 @@
 package com.ll.exam.sbb.answer;
 
-import com.ll.exam.sbb.answer.Answer;
 import com.ll.exam.sbb.base.RepositoryUtil;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.transaction.annotation.Transactional;
 
-public interface AnswerRepository extends JpaRepository<Answer, Integer>, RepositoryUtil {
+public interface AnswerRepository extends JpaRepository<Answer, Long>, RepositoryUtil {
 
     @Transactional
     @Modifying
